{"schema":"decodex/commit/1","summary":"Split sharing and work journal routes","authority":"manual"}
